backend/requirements.txt
django
djangorestframework
markdown
django-filter
django-extensions